Replace ts-rs with typeshare for TypeScript type generation
Switch from ts-rs (derive macro + cargo test export hack) to typeshare
(CLI tool by 1Password) for generating TypeScript interfaces from Rust
structs. This gives us a cleaner generation workflow — typeshare parses
source files directly instead of relying on test side effects.

Key changes:
- `#[derive(TS)] + #[ts(export)]` → `#[typeshare]` attribute
- Per-type .ts files → single generated types.ts
- `cargo test export_bindings` → `typeshare` CLI in scripts
- TodoListResponse counts: i64 → i32 (typeshare doesn't support i64)
- Option<T> now maps to optional fields (T?) instead of T | null

+++ src/models.rs
@@ -1,7 +1,7 @@
 #[cfg(feature = "postgres")]
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
-use ts_rs::TS;
+use typeshare::typeshare;
 #[cfg(feature = "postgres")]
 use uuid::Uuid;
 
@@ -31,9 +31,9 @@ pub struct TodoRow {
     pub updated_at: DbTimestamp,
 }
 
-#[derive(Debug, Clone, Serialize, TS)]
+#[derive(Debug, Clone, Serialize)]
 #[serde(rename_all = "camelCase")]
-#[ts(export)]
+#[typeshare]
 pub struct Todo {
     pub id: String,
     pub title: String,
@@ -42,26 +42,24 @@ pub struct Todo {
     pub updated_at: String,
 }
 
-#[derive(Debug, Clone, Serialize, TS)]
+#[derive(Debug, Clone, Serialize)]
 #[serde(rename_all = "camelCase")]
-#[ts(export)]
+#[typeshare]
 pub struct TodoListResponse {
     pub todos: Vec<Todo>,
-    #[ts(type = "number")]
-    pub active_count: i64,
-    #[ts(type = "number")]
-    pub completed_count: i64,
+    pub active_count: i32,
+    pub completed_count: i32,
 }
 
-#[derive(Debug, Deserialize, TS)]
-#[ts(export)]
+#[derive(Debug, Deserialize)]
+#[typeshare]
 pub struct CreateTodoRequest {
     pub title: String,
 }
 
-#[derive(Debug, Deserialize, TS)]
+#[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
-#[ts(export)]
+#[typeshare]
 pub struct UpdateTodoRequest {
     pub title: Option<String>,
     pub completed: Option<bool>,
